+++ circuits/src/hash/sha256/sha256_chip.rs
@@ -76,7 +76,7 @@ pub const NB_SHA256_ADVICE_COLS: usize = 8;
 /// Number of fixed columns used by the identities of the SHA256 chip.
 pub const NB_SHA256_FIXED_COLS: usize = 2;
 
-const ROUND_CONSTANTS: [u32; 64] = [
+pub(super) const ROUND_CONSTANTS: [u32; 64] = [
     0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
     0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3, 0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174,
     0xe49b69c1, 0xefbe4786, 0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
@@ -87,7 +87,7 @@ const ROUND_CONSTANTS: [u32; 64] = [
     0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208, 0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2,
 ];
 
-const IV: [u32; 8] = [
+pub(super) const IV: [u32; 8] = [
     0x6a09e667, 0xbb67ae85, 0x3c6ef372, 0xa54ff53a, 0x510e527f, 0x9b05688c, 0x1f83d9ab, 0x5be0cd19,
 ];
 
@@ -603,7 +603,7 @@ impl<F: PrimeField> Sha256Chip<F> {
     /// # Panics
     ///
     /// If it does not receive exactly 64 bytes.
-    fn block_from_bytes(
+    pub(super) fn block_from_bytes(
         &self,
         layouter: &mut impl Layouter<F>,
         bytes: &[AssignedByte<F>],
@@ -625,7 +625,7 @@ impl<F: PrimeField> Sha256Chip<F> {
     /// Takes a 512-bits block, represented with 16 `AssignedPlain<32>` words.
     /// Outputs the 64 `AssignedPlain<32>` words Wi from SHA256's message
     /// schedule.
-    fn message_schedule(
+    pub(super) fn message_schedule(
         &self,
         layouter: &mut impl Layouter<F>,
         block: &[AssignedPlain<F, 32>; 16],
@@ -661,7 +661,7 @@ impl<F: PrimeField> Sha256Chip<F> {
     }
 
     /// A compression round. This is called 64 times per block.
-    fn compression_round(
+    pub(super) fn compression_round(
         &self,
         layouter: &mut impl Layouter<F>,
         state: &CompressionState<F>,